feat: stop using scm_info entirely
Assume the source repository is a git repository and use the one that
we get from the context on which bazel runs.

https://bazel.build/docs/user-manual#running-executables describes:

```
The following extra environment variables are also available to the binary:

BUILD_WORKSPACE_DIRECTORY: the root of the workspace where the build was run.
BUILD_WORKING_DIRECTORY: the current working directory where Bazel was run from.
```

which is why we don't need to pass the scm_info arround, by default we support
.git but users can set it to .hg or whatever their SCM has as source directory